+++ CM/CubicEquationSolution/include/solver.h
@@ -3,12 +3,12 @@
 #include <vector>
 
 namespace solver{
-    std::vector<int> solveEquation(float a, float b, float c, float d, float accuracy, double step);
-    int calcDiscriminant(float a, float b, float c, std::vector<float> roots);
-    void calcSquareRoots(float a, float b, float discriminant, std::vector<float> roots);
-    int researchDerivative(float a, float b, float c, std::vector<float> derivativeRoots, bool *isNegative);
-    int calcRoots(int maxRootsCount, float a, float b, float c, float d, double accuracy, double step, std::vector<float> roots);
-    float bisectionMethod(double a, double b, double accuracy, double step);
+    std::vector<double> solveEquation(double a, double b, double c, double d, double accuracy, double step);
+    int calcDiscriminant(double a, double b, double c, std::vector<double> roots);
+    void calcSquareRoots(double a, double b, double discriminant, std::vector<double> roots);
+    int researchDerivative(double a, double b, double c, std::vector<double> derivativeRoots, bool *isNegative);
+    int calcRoots(int maxRootsCount, double a, double b, double c, double d, double accuracy, double step, std::vector<double> roots);
+    double bisectionMethod(double a, double b, double accuracy, double step);
 }
 
 #endif
