+++ CM/CubicEquationSolution/src/solver.cpp
@@ -1,25 +1,25 @@
 #include "solver.h"
 #include <cmath>
 
-std::vector<int> solver::solveEquation(float a, float b, float c, float d, float accuracy, double step) {
-    std::vector<float> derivativeRoots;
-    std::vector<float> solutions;
+std::vector<double> solver::solveEquation(double a, double b, double c, double d, double accuracy, double step) {
+    std::vector<double> derivativeRoots;
+    std::vector<double> solutions;
     bool isNegative = false;
     int maxRootsCount = researchDerivative(a, b, c, derivativeRoots, &isNegative);
     int res = calcRoots(maxRootsCount, a, b, c, d, accuracy, step, solutions);
-    return std::vector<int>();
+    return std::vector<double>();
 }
 
-int solver::calcDiscriminant(float a, float b, float c, std::vector<float> roots) {
+int solver::calcDiscriminant(double a, double b, double c, std::vector<double> roots) {
     return b * b - 4 * a * c;
 }
 
-void solver::calcSquareRoots(float a, float b, float discriminant, std::vector<float> roots) {
+void solver::calcSquareRoots(double a, double b, double discriminant, std::vector<double> roots) {
     roots.push_back((-1 * b + std::sqrt(discriminant)) / (2 * a));
     roots.push_back((-1 * b - std::sqrt(discriminant)) / (2 * a));
 }
 
-int solver::researchDerivative(float a, float b, float c, std::vector<float> derivativeRoots, bool* isNegative) {
+int solver::researchDerivative(double a, double b, double c, std::vector<double> derivativeRoots, bool* isNegative) {
     a *= 3;
     b *= 2;
     int discriminant = calcDiscriminant(a, b, c, derivativeRoots);
@@ -35,9 +35,9 @@ int solver::researchDerivative(float a, float b, float c, std::vector<float> der
     return 0;
 }
 
-int solver::calcRoots(int maxRootsCount, float a, float b, float c, float d, double accuracy, double step, std::vector<float> roots) {
+int solver::calcRoots(int maxRootsCount, double a, double b, double c, double d, double accuracy, double step, std::vector<double> roots) {
     if (maxRootsCount == 1) {
-        float start, finish;
+        double start, finish;
         // функция убывает, т.к. D < 0 и в 3ax^2 + 2bx + c, c < 0.
         if (c < 0) {
             if (d > 0) {
@@ -58,6 +58,6 @@ int solver::calcRoots(int maxRootsCount, float a, float b, float c, float d, dou
     }
 }
 
-float solver::bisectionMethod(double a, double b, double accuracy, double step) {
+double solver::bisectionMethod(double a, double b, double accuracy, double step) {
 
 }
